requirements.txt
# Dermatology Benchmark Dependencies
# Python 3.8+

# Google Gemini API (NEW SDK)
google-genai>=0.2.0

# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
